+++ buttons/buttons_search_by_distance.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//12/03/25
+//05/08/25
 
 include('..\\helpers\\buttons_xxx.js');
 /* global getUniquePrefix:readable, buttonsBar:readable, addButton:readable, ThemedButton:readable */
@@ -15,7 +15,9 @@ include('..\\helpers\\helpers_xxx_UI.js');
 include('..\\helpers\\helpers_xxx_properties.js');
 /* global setProperties:readable, getPropertiesPairs:readable */
 include('..\\main\\search_by_distance\\search_by_distance.js'); // Load after buttons_xxx.js so properties are only set once
-/* global SearchByDistance_properties:readable, music_graph_descriptors:readable, sbd:readable, searchByDistance:readable, updateCache:readable, findStyleGenresMissingGraphCheck:readable, testBaseTags:readable */
+/* global SearchByDistance_properties:readable, sbd:readable, updateCache:readable, findStyleGenresMissingGraphCheck:readable, testBaseTags:readable */
+include('helpers\\buttons_sbd_menu_presets.js'); // Button menu
+/* global choosePresetMenu:readable */
 var version = sbd.version; // NOSONAR[global]
 
 try { window.DefineScript('Search by Distance Buttons', { author: 'regorxxx', version, features: { drag_n_drop: false } }); } catch (e) { /* May be loaded along other buttons */ } // eslint-disable-line no-unused-vars
@@ -72,99 +74,25 @@ setProperties(newButtonsProperties, prefix, 0); //This sets all the panel proper
 */
 
 addButton({
-	'Search by Distance nearest tracks': new ThemedButton({
-		coordinates: { x: 0, y: 0, w: _gr.CalcTextWidth('Nearest Tracks', _gdiFont(globFonts.button.name, globFonts.button.size * buttonsBar.config.scale)) + 25 * _scale(1, false) / _scale(buttonsBar.config.scale), h: 22 },
-		text: 'Nearest Tracks',
-		func: function (mask) {
-			if (mask === MK_SHIFT) {
-				settingsMenu(this, true, ['buttons_search_by_distance.js'], { bAdvTitle: { popup: globRegExp.title.desc } }).btn_up(this.currX, this.currY + this.currH);
-			} else {
-				// Mix with only nearest tracks
-				const tags = JSON.parse(this.buttonsProperties.tags[1]);
-				tags.genre.weight = 15; tags.style.weight = 10; tags.mood.weight = 5; tags.key.weight = 10; tags.date.weight = 25; tags.bpm.weight = 5;
-				tags.date.range = 15; tags.bpm.range = 25;
-				tags.dynGenre.weight = 25; tags.dynGenre.range = 1;
-				const defArgs = { tags, bRandomPick: true, bHarmonicMixDoublePass: true, properties: this.buttonsProperties };
-				const args = { ...defArgs, minScoreFilter: 65, scoreFilter: 70, graphDistance: music_graph_descriptors.intra_supergenre / 2 };
-				searchByDistance(args);
-			}
-		},
-		description: buttonTooltipSbd,
-		prefix, buttonsProperties: newButtonsProperties,
-		icon: chars.wand,
-		variables: { descriptionText: 'Random mix with only nearest tracks' }
-	}),
-	'Search by Distance similar tracks': new ThemedButton({
+	'Search by Distance presets': new ThemedButton({
 		coordinates: { x: 0, y: 0, w: _gr.CalcTextWidth('Similar Tracks', _gdiFont(globFonts.button.name, globFonts.button.size * buttonsBar.config.scale)) + 25 * _scale(1, false) / _scale(buttonsBar.config.scale), h: 22 },
 		text: 'Similar Tracks',
 		func: function (mask) {
 			if (mask === MK_SHIFT) {
 				settingsMenu(this, true, ['buttons_search_by_distance.js'], { bAdvTitle: { popup: globRegExp.title.desc } }).btn_up(this.currX, this.currY + this.currH);
 			} else {
-				// Mix a bit varied on styles/genres most from the same decade
-				const tags = JSON.parse(this.buttonsProperties.tags[1]);
-				tags.genre.weight = 15; tags.style.weight = 10; tags.mood.weight = 5; tags.key.weight = 5; tags.date.weight = 25; tags.bpm.weight = 5;
-				tags.date.range = 15; tags.bpm.range = 25;
-				tags.dynGenre.weight = 10; tags.dynGenre.range = 1;
-				const defArgs = { tags, bRandomPick: true, bHarmonicMixDoublePass: true, properties: this.buttonsProperties };
-				const args = { ...defArgs, minScoreFilter: 55, scoreFilter: 60, graphDistance: music_graph_descriptors.cluster };
-				searchByDistance(args);
+				choosePresetMenu(this).btn_up(this.currX, this.currY + this.currH);
 			}
 		},
 		description: buttonTooltipSbd,
 		prefix, buttonsProperties: newButtonsProperties,
-		icon: chars.wand, variables: { descriptionText: 'Random mix a bit varied on styles (but similar genre), most tracks within a decade' }
-	}),
-	'Search by Distance similar genres': new ThemedButton({
-		coordinates: { x: 0, y: 0, w: _gr.CalcTextWidth('Similar Genres', _gdiFont(globFonts.button.name, globFonts.button.size * buttonsBar.config.scale)) + 25 * _scale(1, false) / _scale(buttonsBar.config.scale), h: 22 },
-		text: 'Similar Genres',
-		func: function (mask) {
-			if (mask === MK_SHIFT) {
-				settingsMenu(this, true, ['buttons_search_by_distance.js'], { bAdvTitle: { popup: globRegExp.title.desc } }).btn_up(this.currX, this.currY + this.currH);
-			} else {
-				// Mix even more varied on styles/genres most from the same decade
-				const tags = JSON.parse(this.buttonsProperties.tags[1]);
-				tags.genre.weight = 0; tags.style.weight = 5; tags.mood.weight = 15; tags.key.weight = 10; tags.date.weight = 25; tags.bpm.weight = 5;
-				tags.date.range = 15; tags.bpm.range = 25;
-				tags.dynGenre.weight = 10; tags.dynGenre.range = 2;
-				const defArgs = { tags, bRandomPick: true, bHarmonicMixDoublePass: true, properties: this.buttonsProperties };
-				const args = { ...defArgs, minScoreFilter: 55, scoreFilter: 60, graphDistance: music_graph_descriptors.intra_supergenre * 3 / 2 };
-				searchByDistance(args);
-			}
-		},
-		description: buttonTooltipSbd,
-		prefix, buttonsProperties: newButtonsProperties,
-		icon: chars.wand,
-		variables: { descriptionText: 'Random mix even more varied on styles/genres, most tracks within a decade' }
-	}),
-	'Search by Distance similar mood': new ThemedButton({
-		coordinates: { x: 0, y: 0, w: _gr.CalcTextWidth('Similar Mood', _gdiFont(globFonts.button.name, globFonts.button.size * buttonsBar.config.scale)) + 25 * _scale(1, false) / _scale(buttonsBar.config.scale), h: 22 },
-		text: 'Similar Mood',
-		func: function (mask) {
-			if (mask === MK_SHIFT) {
-				settingsMenu(this, true, ['buttons_search_by_distance.js'], { bAdvTitle: { popup: globRegExp.title.desc } }).btn_up(this.currX, this.currY + this.currH);
-			} else {
-				// Mix with different genres but same mood from any date
-				const tags = JSON.parse(this.buttonsProperties.tags[1]);
-				tags.genre.weight = 0; tags.style.weight = 5; tags.mood.weight = 15; tags.key.weight = 10; tags.date.weight = 0; tags.bpm.weight = 5;
-				tags.date.range = 0; tags.bpm.range = 25;
-				tags.dynGenre.weight = 5; tags.dynGenre.range = 4;
-				const defArgs = { tags, bRandomPick: true, bHarmonicMixDoublePass: true, properties: this.buttonsProperties };
-				const args = { ...defArgs, minScoreFilter: 45, scoreFilter: 50, graphDistance: music_graph_descriptors.intra_supergenre * 4 };
-				searchByDistance(args);
-			}
-		},
-		description: buttonTooltipSbd,
-		prefix, buttonsProperties: newButtonsProperties,
-		icon: chars.wand,
-		variables: { descriptionText: 'Random mix with different genres but same mood from any date' },
-		update: { scriptName: 'Search-by-Distance-SMP', version }
-	}),
+		icon: chars.wand
+	})
 });
 
 // Helper
 function buttonTooltipSbd() {
-	let info = this.descriptionText + ':';
+	let info = 'Search similar tracks by acoustic-folksonomy models:';
 	info += '\nMethod:\t' + this.buttonsProperties.method[1];
 	const sort = (
 		(this.buttonsProperties.bSmartShuffle[1] ? 'Smart Shuffle' : '')